aplicacion_cultura_api_rest/requirements.txt
mysqlclient==2.1.1
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
